+++ fleetshard/pkg/fleetmanager/auth.go
@@ -2,14 +2,10 @@ package fleetmanager
 
 import (
 	_ "embed"
-	"errors"
 	"fmt"
 	"github.com/golang/glog"
-	sdk "github.com/openshift-online/ocm-sdk-go"
 	"net/http"
-	"os"
 	"strings"
-	"time"
 )
 
 // AuthType represents the supported authentication types for the client.
@@ -52,54 +48,8 @@ type Auth interface {
 func NewAuth(t AuthType) (Auth, error) {
 	switch t {
 	default:
-		return newOcmAuth(os.Getenv("OCM_TOKEN"))
+		return newOcmAuth()
 	}
-}
-
-type ocmAuth struct {
-	conn *sdk.Connection
-}
-
-func newOcmAuth(initialToken string) (*ocmAuth, error) {
-	if initialToken == "" {
-		return nil, errors.New("empty ocm token")
-	}
-
-	l, err := sdk.NewGlogLoggerBuilder().Build()
-	if err != nil {
-		return nil, err
-	}
-
-	builder := sdk.NewConnectionBuilder().
-		Client("cloud-services", "").
-		Tokens(initialToken).
-		Logger(l)
-
-	// Check if the connection can be established and tokens can be retrieved.
-	conn, err := builder.Build()
-	if err != nil {
-		return nil, err
-	}
-	_, _, err = conn.Tokens()
-	if err != nil {
-		return nil, err
-	}
-
-	return &ocmAuth{
-		conn: conn,
-	}, nil
-}
-
-func (o *ocmAuth) AddAuth(req *http.Request) error {
-	// This will only do an external request iff the current access token of the connection has an expiration time
-	// lower than 1 minute.
-	access, _, err := o.conn.TokensContext(req.Context(), 1*time.Minute)
-	if err != nil {
-		return err
-	}
-
-	setBearer(req, access)
-	return nil
 }
 
 type noAuth struct{}
@@ -110,10 +60,7 @@ func (n noAuth) AddAuth(_ *http.Request) error {
 
 // setBearer is a helper to set a bearer token as authorization header on the http.Request.
 func setBearer(req *http.Request, token string) {
-	// Do not attempt to modify any existing authorization headers.
-	if req.Header.Get("Authorization") == "" {
-		req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", token))
-	}
+	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", token))
 }
 
 // getAllAuthTypes is a helper used within logging to list the possible values for auth types.
