--- fleetshard/pkg/fleetmanager/static-token
@@ -1,1 +1,0 @@
-eyJrdHkiOiJSU0EiLCJlIjoiQVFBQiIsInVzZSI6InNpZyIsImtpZCI6ImwzLTU2VmdRTFBCY3NkNERKdHBiYVQ5QllJeDFTeEtXOElsS0FTUHdvQlEiLCJhbGciOiJSUzI1NiIsIm4iOiIzNXAwbGp3cmYxNkQycFN6VTdQX3I2bWgwUkpJSXZTdEJjNjJ5ZlhMRkpONWhkMWs1NHFKZFJ0OXZIdmdVYXc1UjVQVlNudFNHZDVNZEZMMlFjSXlLRkZhdnhYdXV2MURMeGgtZUtDZHg0eDMyWFFfSTZqQzcxcWpsVXdYclRSdmw1SVllRFZNYmwyYXFxdnF5N0lGc1RJbEVVQzN0eWNwN2ZOdzNQZk9ZVHlBa0VfbTNsc0M4bE5zMkxRQzNPcXBNb2ZhSHdsVTlTcU83U2o5NVFVR2ZjSlpZT24tdGl2dmNnNlh0OWlpMW5FT2Q3TlhjaWdYN2tLUTJFUjlFLWx0QXZiTnV2QmtHa2VOMU9zNEdiUkpSRU5fOWxyNTZuaWFERk93VlU5ZEpIMi0wLU84bFFuOXBSanF3SlhpbEx4a3J2WXZFR0xZc1RUWWctb29FZm83Z1EifQ.eyJpYXQiOjE2NTYyOTM5MjcsImp0aSI6IjAzZjBiZDdiLTAwM2YtNDNjYy1hMTY5LWZjNGE3YTU4MjY1NSIsImlzcyI6Imh0dHBzOi8vc3NvLnJlZGhhdC5jb20vYXV0aC9yZWFsbXMvcmVkaGF0LWV4dGVybmFsIiwiZXhwIjo3OTU1MTEyMTQyLCJhdWQiOiJjbG91ZC1zZXJ2aWNlcyIsInN1YiI6ImY6NTI4ZDc2ZmYtZjcwOC00M2VkLThjZDUtZmUxNmY0ZmUwY2U2OnJoYWNzLW1zLXRlc3RAcmVkaGF0LmNvbSIsInR5cCI6IkJlYXJlciIsImF6cCI6ImNsb3VkLXNlcnZpY2VzIiwibm9uY2UiOiJlYjE0ODVkNi04MzUxLTQ0OTQtOTc4ZC04YjYyYmViZTEyNmIiLCJzZXNzaW9uX3N0YXRlIjoiMjA4NzNkMTItYWFlOC00ZDNiLTljMTQtNDRhNWEyNTNjMzY3IiwiYWNyIjoiMCIsInJlYWxtX2FjY2VzcyI6eyJyb2xlcyI6WyJhdXRoZW50aWNhdGVkIiwiZmxlZXQtbWFuYWdlci1hZG1pbi1mdWxsIl19LCJzY29wZSI6Im9wZW5pZCBpYW0uY2xpZW50cy5zZXJ2aWNlX2FjY291bnRzIG9mZmxpbmVfYWNjZXNzIiwic2lkIjoiMjA4NzNkMTItYWFlOC00ZDNiLTljMTQtNDRhNWEyNTNjMzY3IiwiYWNjb3VudF9udW1iZXIiOiIxMjM0NTY3IiwiaXNfaW50ZXJuYWwiOnRydWUsImlzX2FjdGl2ZSI6dHJ1ZSwibGFzdF9uYW1lIjoiVGVzdCIsInByZWZlcnJlZF91c2VybmFtZSI6InJoYWNzLW1zLXRlc3RAcmVkaGF0LmNvbSIsInR5cGUiOiJVc2VyIiwibG9jYWxlIjoiZW5fdXMiLCJpc19vcmdfYWRtaW4iOnRydWUsImFjY291bnRfaWQiOiIxMjM0NTY3OCIsInVzZXJfaWQiOiIxMjM0NTY3OCIsIm9yZ19pZCI6IjEyMzQ1Njc4IiwiZmlyc3RfbmFtZSI6IlRlc3QiLCJlbWFpbCI6InJoYWNzLW1zLXRlc3RAcmVkaGF0LmNvbSIsInVzZXJuYW1lIjoicmhhY3MtbXMtdGVzdEByZWRoYXQuY29tIn0.OlX7HrtPhlm0AZF_qj7IGemAUVbJX4hv8voNGZDIwK0vs2CoMzKiGEPqy6ds3cujS4zw8-GN35sKE2FzYuqy6fJpzG_v65P7l1b2jxri3a01CXbwUQS3c-_B5qyUMgfsA7Jcr57kwyc2JOHeITUibJ3mNa9J8UEVnlD5qLcoLC8gIH465qgaIzo4r_jry47lzSAU8ZI_AcQCIX8qkf4ZuHSR6gXYd-cHGMxfsHCdg3JGU4HtKVJjJq8lHdsI6KQuyC8AuVew02awoovJ5jHl2IvoDZw3vrsLBa4YwTrxI5o1G7WmaDAxV1rw7ZzGYA91XCJm6VZ2h36gRqvNPRbGuQ
